+++ src/app/item-picker/ItemPicker.tsx
@@ -19,6 +19,7 @@ import { itemSortOrderSelector } from '../settings/item-sort';
 import classNames from 'classnames';
 import { t } from 'i18next';
 import './ItemPicker.scss';
+import { setSetting } from '../settings/actions';
 
 type ProvidedProps = ItemPickerState & {
   onSheetClosed(): void;
@@ -30,6 +31,7 @@ interface StoreProps {
   filters: SearchFilters;
   itemSortOrder: string[];
   isPhonePortrait: boolean;
+  preferEquip: boolean;
 }
 
 function mapStateToProps(): MapStateToProps<StoreProps, ProvidedProps, RootState> {
@@ -45,11 +47,17 @@ function mapStateToProps(): MapStateToProps<StoreProps, ProvidedProps, RootState
     searchConfig: searchConfigSelector(state),
     filters: searchFiltersConfigSelector(state),
     itemSortOrder: itemSortOrderSelector(state),
-    isPhonePortrait: state.shell.isPhonePortrait
+    isPhonePortrait: state.shell.isPhonePortrait,
+    preferEquip: state.settings.itemPickerEquip
   });
 }
 
-type Props = ProvidedProps & StoreProps;
+const mapDispatchToProps = {
+  setSetting
+};
+type DispatchProps = typeof mapDispatchToProps;
+
+type Props = ProvidedProps & StoreProps & DispatchProps;
 
 interface State {
   query: string;
@@ -58,7 +66,7 @@ interface State {
 }
 
 class ItemPicker extends React.Component<Props, State> {
-  state: State = { query: '', equip: true };
+  state: State = { query: '', equip: this.props.preferEquip };
   private itemContainer = React.createRef<HTMLDivElement>();
   private filterInput = React.createRef<SearchFilterInput>();
 
@@ -84,7 +92,7 @@ class ItemPicker extends React.Component<Props, State> {
 
     const header = (
       <div>
-        <h1>{prompt || t('ItemPicker.ChooseItem')}</h1>
+        <h1 className="destiny">{prompt || t('ItemPicker.ChooseItem')}</h1>
         <div className="item-picker-search">
           <SearchFilterInput
             ref={this.filterInput}
@@ -143,8 +151,17 @@ class ItemPicker extends React.Component<Props, State> {
     this.props.onSheetClosed();
   };
 
-  private setEquip = () => this.setState({ equip: true });
-  private setStore = () => this.setState({ equip: false });
+  private setEquip = () => {
+    this.setState({ equip: true });
+    this.props.setSetting('itemPickerEquip', true);
+  };
+  private setStore = () => {
+    this.setState({ equip: false });
+    this.props.setSetting('itemPickerEquip', false);
+  };
 }
 
-export default connect<StoreProps>(mapStateToProps)(ItemPicker);
+export default connect<StoreProps, DispatchProps>(
+  mapStateToProps,
+  mapDispatchToProps
+)(ItemPicker);